+++ aurora-springboot/src/main/java/com/aurora/service/impl/CategoryServiceImpl.java
@@ -42,7 +42,7 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     @Override
     public PageResultDTO<CategoryAdminDTO> listCategoriesAdmin(ConditionVO conditionVO) {
         Integer count = categoryMapper.selectCount(new LambdaQueryWrapper<Category>()
-                .like(StringUtils.isNotBlank(conditionVO.getKeywords()), Category::getCategoryName, conditionVO.getKeywords()));
+                .like(StringUtils.isNotBlank(conditionVO.getKeywords()), Category::getCategoryName, conditionVO.getKeywords())).intValue();
         if (count == 0) {
             return new PageResultDTO<>();
         }
@@ -63,7 +63,7 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     @Override
     public void deleteCategories(List<Integer> categoryIds) {
         Integer count = articleMapper.selectCount(new LambdaQueryWrapper<Article>()
-                .in(Article::getCategoryId, categoryIds));
+                .in(Article::getCategoryId, categoryIds)).intValue();
         if (count > 0) {
             throw new BizException("删除失败，该分类下存在文章");
         }
