+++ aurora-springboot/src/main/java/com/aurora/service/impl/ResourceServiceImpl.java
@@ -95,7 +95,7 @@ public class ResourceServiceImpl extends ServiceImpl<ResourceMapper, Resource> i
     @Override
     public void deleteResource(Integer resourceId) {
         Integer count = roleResourceMapper.selectCount(new LambdaQueryWrapper<RoleResource>()
-                .eq(RoleResource::getResourceId, resourceId));
+                .eq(RoleResource::getResourceId, resourceId)).intValue();
         if (count > 0) {
             throw new BizException("该资源下存在角色");
         }
