+++ aurora-springboot/src/main/java/com/aurora/service/impl/RoleServiceImpl.java
@@ -62,7 +62,7 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
     public PageResultDTO<RoleDTO> listRoles(ConditionVO conditionVO) {
         LambdaQueryWrapper<Role> queryWrapper = new LambdaQueryWrapper<Role>()
                 .like(StringUtils.isNotBlank(conditionVO.getKeywords()), Role::getRoleName, conditionVO.getKeywords());
-        CompletableFuture<Integer> asyncCount = CompletableFuture.supplyAsync(() -> roleMapper.selectCount(queryWrapper));
+        CompletableFuture<Integer> asyncCount = CompletableFuture.supplyAsync(() -> roleMapper.selectCount(queryWrapper).intValue());
         List<RoleDTO> roleDTOs = roleMapper.listRoles(PageUtil.getLimitCurrent(), PageUtil.getSize(), conditionVO);
         return new PageResultDTO<>(roleDTOs, asyncCount.get());
     }
@@ -113,7 +113,7 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
     @Override
     public void deleteRoles(List<Integer> roleIdList) {
         Integer count = userRoleMapper.selectCount(new LambdaQueryWrapper<UserRole>()
-                .in(UserRole::getRoleId, roleIdList));
+                .in(UserRole::getRoleId, roleIdList)).intValue();
         if (count > 0) {
             throw new BizException("该角色下存在用户");
         }
