+++ aurora-springboot/src/main/java/com/aurora/service/impl/TagServiceImpl.java
@@ -47,7 +47,7 @@ public class TagServiceImpl extends ServiceImpl<TagMapper, Tag> implements TagSe
     @Override
     public PageResultDTO<TagAdminDTO> listTagsAdmin(ConditionVO conditionVO) {
         Integer count = tagMapper.selectCount(new LambdaQueryWrapper<Tag>()
-                .like(StringUtils.isNotBlank(conditionVO.getKeywords()), Tag::getTagName, conditionVO.getKeywords()));
+                .like(StringUtils.isNotBlank(conditionVO.getKeywords()), Tag::getTagName, conditionVO.getKeywords())).intValue();
         if (count == 0) {
             return new PageResultDTO<>();
         }
@@ -79,7 +79,7 @@ public class TagServiceImpl extends ServiceImpl<TagMapper, Tag> implements TagSe
     @Override
     public void deleteTag(List<Integer> tagIds) {
         Integer count = articleTagMapper.selectCount(new LambdaQueryWrapper<ArticleTag>()
-                .in(ArticleTag::getTagId, tagIds));
+                .in(ArticleTag::getTagId, tagIds)).intValue();
         if (count > 0) {
             throw new BizException("删除失败，该标签下存在文章");
         }
